+++ familyhelper-clannad-impl/src/main/java/com/dwarfeng/familyhelper/clannad/impl/util/FtpConstants.java
@@ -1,0 +1,16 @@
+package com.dwarfeng.familyhelper.clannad.impl.util;
+
+/**
+ * FTP 常量。
+ *
+ * @author DwArFeng
+ * @since 1.1.4
+ */
+public final class FtpConstants {
+
+    public static final String PATH_AVATAR = "avatar";
+
+    private FtpConstants() {
+        throw new IllegalStateException("禁止实例化");
+    }
+}